Separate Types file & update imports
