+++ modules/server/src/mapping/utils/esSearch.ts
@@ -1,6 +1,6 @@
 import { type RequestParams } from '@elastic/elasticsearch';
 
-import { type SearchClient } from '#searchClient/index.js';
+import { type SearchClient } from '#searchClient/types.js';
 
 export default (esClient: SearchClient) => async (params: RequestParams.Search) => {
 	return esClient.search(params)?.body;
